# improve AudioContext#sched()