benchmark

before

```
Mozilla/5.0 (Macintosh; Intel Mac OS X 10.11; rv:45.0) Gecko/20100101 Firefox/45.0
web-audio-engine x 6.36 ops/sec ±1.03% (34 runs sampled)
native Web Audio API x 839 ops/sec ±3.36% (50 runs sampled)
Fastest is native Web Audio API
```

after

```
Mozilla/5.0 (Macintosh; Intel Mac OS X 10.11; rv:45.0) Gecko/20100101 Firefox/45.0
web-audio-engine x 179 ops/sec ±12.12% (53 runs sampled)
native Web Audio API x 838 ops/sec ±2.33% (53 runs sampled)
Fastest is native Web Audio API
```

6.36opts -> 179 opts